Questher v3 - Week 3 Implementation/requirements.txt
gradio>=4.0.0
openai>=1.0.0
python-dotenv>=1.0.0
matplotlib>=3.7.0
numpy>=1.24.0